fix: converted the table into a component
